TEMPLATE H1
added a media query to the h1 to center the text on mobile devices. I
added a class page-content so I could target the h1 in just the content
section. Not sure, but maybe I’ll be using the H1 somewhere else and I
won’t want it to center. I can take this class out once the sites built
and it’s clear where H1 is being used.