TEMPLATE BUGS
auto-size tag wasn’t closed - this isn’t working anyway. I’m trying to
get the blue at the bottom of the page to extend to end of the browser
no matter the size of the page.

I uncommented out the html tags - I’m not sure what these are doing.
This might have come from initlizr. Need to check this.