LEFT NAV
tried to get a line on top of the nav that I liked, but I can’t get the
line and the text to line up on the right side. The text has some
padding on it to accommodate the bootstrap light grey rollover.

I also made my padding and margin library more complete. added comments
in to make it more legible