Lightbox
implementing a bootstrap modal for a lightbox. I was over-writing all
modals with a fix to center the separators on the main menu modal. I
stopped doing that and created a new class “custom-modal-center-fix”

I have a first pass at the bootstrap lightbox on
salmon-recovery-plans.php but it’s taking up the entire viewport. I am
going to try and make it take up less.