JONSTOOL
Turned on the modalManager function - it’s called from
project[‘marker’] listener. it works.

I changed the coordinates of the hit area poly on the markers. Need to
research a little to make sure I did this write. Look up the shape for
the coordinates I gave the markers. I can’t see it, but the hit areas
feels about right.