PRESS
for testing I detached from template - this works but isn’t in the
template.

+ fixed bug in nav code that was throwing an error when subNavSelected
wasn’t present. This was breaking my news release code but was
unrelated. I fixed it by adding a variable into the html for subnav
when it isn’t present. I’ll need to add this to all the pages.

+ the press json file is working! the new code is in the custom.js file.